// src/app/models/apache-health-information.ts
import { MotorResponse } from 'src/app/models/motor-response';
import { VerbalResponse } from 'src/app/models/verbal-response';
import { Origin } from 'src/app/models/origin';
import { AdmissionDiagnosis } from 'src/app/models/admission-diagnosis';
import { CrfHd } from 'src/app/models/crf-hd';
import { Sed } from 'src/app/services/apache-4-calculator/sed';
import { VentilationMode } from 'src/app/models/ventilation-mode';

export interface ApacheHealthInformation {
  age: number;
  temperature: number;
  map: number;
  heartRate: number;
  respirationRate: number;
  fio: number;
  oxy: number;
  pco: number;
  arterialPh: number;
  nas: number;
  urea: number;
  creatine: number;
  urine: number;
  bsl: number;
  albumin: number;
  bilirubin: number;
  hto: number;
  wbc: number;
  gcEyes: number;
  gcVerbal: VerbalResponse;
  gcMotor: MotorResponse;
  respiratoryQuotient: number;
  atmosphericPressure: number;
  los: number;
  origin: Origin;
  admissionDiagnosis: AdmissionDiagnosis;
  system: number;
  diagnosis: number;
  readmission: boolean;
  emergencySurgery: boolean;
  thrombolysis: number;
  crfHd: CrfHd;
  sed: Sed;
  isAidChecked: boolean;
  isHepChecked: boolean;
  isLymChecked: boolean;
  isMetChecked: boolean;
  isLeuChecked: boolean;
  isImmChecked: boolean;
  isCirChecked: boolean;
  ventilationMode: VentilationMode;
}